requirements.txt
python 3.13
Django-5.1.7

